03 May Harshit by Coomitn

+++ app/src/main/java/com/tech/twe_tutorside/model/HourlyDetailsModel.java
@@ -7,22 +7,22 @@ public class HourlyDetailsModel {
 
     @SerializedName("service_charge")
     @Expose
-    private Integer serviceCharge;
+    private Double serviceCharge;
     @SerializedName("discount_to_student")
     @Expose
     private String discountToStudent;
     @SerializedName("teachers_earning")
     @Expose
-    private Integer teachersEarning;
+    private Double teachersEarning;
     @SerializedName("student_price")
     @Expose
     private String studentPrice;
 
-    public Integer getServiceCharge() {
+    public Double getServiceCharge() {
         return serviceCharge;
     }
 
-    public void setServiceCharge(Integer serviceCharge) {
+    public void setServiceCharge(Double serviceCharge) {
         this.serviceCharge = serviceCharge;
     }
 
@@ -34,11 +34,11 @@ public class HourlyDetailsModel {
         this.discountToStudent = discountToStudent;
     }
 
-    public Integer getTeachersEarning() {
+    public Double getTeachersEarning() {
         return teachersEarning;
     }
 
-    public void setTeachersEarning(Integer teachersEarning) {
+    public void setTeachersEarning(Double teachersEarning) {
         this.teachersEarning = teachersEarning;
     }
